Consolidate V2 component files (#1902)
* Move a bunch of V2 project files around

* More

* Move more
