+++ src/components/v2/V2Project/ProjectActivity/index.tsx
@@ -21,7 +21,7 @@ import { RedeemEvent } from 'models/subgraph-entities/vX/redeem-event'
 import { useContext, useMemo, useState } from 'react'
 import { WhereConfig } from 'utils/graph'
 
-import V2DownloadActivityModal from '../V2DownloadActivityModal'
+import V2DownloadActivityModal from '../modals/V2DownloadActivityModal'
 import DeployETHERC20ProjectPayerEventElem from './eventElems/DeployETHERC20ProjectPayerEventElem'
 import DistributePayoutsElem from './eventElems/DistributePayoutsElem'
 import DistributeReservedTokensEventElem from './eventElems/DistributeReservedTokensElem'
