+++ src/components/v2/V2Project/V2ProjectSettings/pages/V1V2TokenMigrationSettingsPage/V1V2TokenMigrationSettingsPage.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { V1TokenMigrationSetupSection } from '../../V2ProjectToolsDrawer/V1TokenMigrationSetupSection'
+import { V1TokenMigrationSetupSection } from './V1TokenMigrationSetupSection'
 
 export function V1V2TokenMigrationSettingsPage() {
   return <V1TokenMigrationSetupSection />
