+++ src/components/v2/V2Project/V2ProjectSettings/pages/V1V2TokenMigrationSettingsPage/index.ts
@@ -1,0 +1,1 @@
+export { V1V2TokenMigrationSettingsPage } from './V1V2TokenMigrationSettingsPage'
