+++ src/components/v2/V2Project/V2ProjectSettings/pages/V2ArchiveProjectSettingsPage/V2ArchiveProjectSettingsPage.tsx
@@ -1,8 +1,8 @@
 import { useEditV2ProjectDetailsTx } from 'hooks/v2/transactor/EditV2ProjectDetailsTx'
 import React from 'react'
-import ArchiveV2Project from '../../ArchiveV2Project'
+import { V2ArchiveProject } from './ArchiveV2Project'
 
 export function V2ArchiveProjectSettingsPage() {
   const editV2ProjectDetailsTx = useEditV2ProjectDetailsTx()
-  return <ArchiveV2Project editV2ProjectDetailsTx={editV2ProjectDetailsTx} />
+  return <V2ArchiveProject editV2ProjectDetailsTx={editV2ProjectDetailsTx} />
 }
