+++ src/components/v2/V2Project/V2ProjectSettings/pages/V2ArchiveProjectSettingsPage/index.ts
@@ -1,0 +1,1 @@
+export { V2ArchiveProjectSettingsPage } from './V2ArchiveProjectSettingsPage'
