+++ src/components/v2/V2Project/V2ProjectSettings/pages/V2PayoutsSettingsPage/index.ts
@@ -1,0 +1,1 @@
+export { V2PayoutsSettingsPage } from './V2PayoutsSettingsPage'
