+++ src/components/v2/V2Project/V2ProjectSettings/pages/V2ProjectReconfigureFundingCycleSettingsPage/index.ts
@@ -1,0 +1,1 @@
+export { V2ReconfigureFundingCycleSettingsPage } from './V2ReconfigureFundingCycleSettingsPage'
