+++ src/components/v2/V2Project/V2ProjectSettings/pages/V2ReservedTokensSettingsPage/index.ts
@@ -1,0 +1,1 @@
+export { V2ReservedTokensSettingsPage } from './V2ReservedTokensSettingsPage'
