+++ src/components/v2/V2Project/V2ProjectToolsDrawer/PaymentAddressSection/index.ts
@@ -1,0 +1,1 @@
+export { PaymentAddressSection } from './PaymentAddressSection'
