+++ src/components/v2/V2Project/index.tsx
@@ -29,10 +29,10 @@ import { v2ProjectRoute } from 'utils/routes'
 
 import { V2_PROJECT_IDS } from 'constants/v2/projectIds'
 import { RelaunchFundingCycleBanner } from './banners/RelaunchFundingCycleBanner'
-import NewDeployModal from './NewDeployModal'
+import NewDeployModal from './modals/NewDeployModal'
+import { V2BalancesModal } from './modals/V2BalancesModal/V2BalancesModal'
 import ProjectActivity from './ProjectActivity'
 import TreasuryStats from './TreasuryStats'
-import { V2BalancesModal } from './V2BalancesModal/V2BalancesModal'
 import V2FundingCycleSection from './V2FundingCycleSection'
 import V2ManageTokensSection from './V2ManageTokensSection'
 import V2PayButton from './V2PayButton'
