+++ src/components/v2/V2Project/modals/LaunchProjectPayerModal/LaunchProjectPayerModal.tsx
@@ -35,7 +35,7 @@ export interface AdvancedOptionsFields {
   preferClaimed: boolean
 }
 
-export default function LaunchProjectPayerModal({
+export function LaunchProjectPayerModal({
   visible,
   onClose,
   useDeployProjectPayerTx,
