+++ src/components/v2/V2Project/modals/LaunchProjectPayerModal/index.ts
@@ -1,0 +1,1 @@
+export { LaunchProjectPayerModal } from './LaunchProjectPayerModal'
