+++ src/components/v2/V2Project/modals/V2BalancesModal/V2BalancesModal.tsx
@@ -4,7 +4,7 @@ import { t, Trans } from '@lingui/macro'
 import { Button, Modal, Space } from 'antd'
 import { useForm } from 'antd/lib/form/Form'
 import ERC20TokenBalance from 'components/v1/shared/ERC20TokenBalance'
-import { V2ProjectTokenBalance } from 'components/v2/shared/V2ProjectTokenBalance'
+import { V2ProjectTokenBalance } from 'components/v2/V2Project/modals/V2BalancesModal/V2ProjectTokenBalance'
 import { V2ProjectContext } from 'contexts/v2/projectContext'
 import { TransactorInstance } from 'hooks/Transactor'
 import { ProjectMetadataV4 } from 'models/project-metadata'
