+++ src/components/v2/shared/FundingCycleConfigurationDrawers/FundingDrawer.tsx
@@ -2,8 +2,7 @@ import { t } from '@lingui/macro'
 import FundingCycleDrawer from 'components/v2/shared/FundingCycleConfigurationDrawers/FundingCycleDrawer'
 import { useFundingCycleDrawer } from 'components/v2/shared/FundingCycleConfigurationDrawers/useFundingCycleDrawer'
 import UnsavedChangesModal from 'components/v2/shared/UnsavedChangesModal'
-
-import FundingForm from '../../../../pages/create/forms/FundingForm'
+import FundingForm from 'pages/create/forms/FundingForm'
 
 export default function FundingDrawer({
   visible,
