+++ src/components/v2/shared/FundingCycleConfigurationDrawers/TokenDrawer.tsx
@@ -3,8 +3,7 @@ import Callout from 'components/Callout'
 import FundingCycleDrawer from 'components/v2/shared/FundingCycleConfigurationDrawers/FundingCycleDrawer'
 import { useFundingCycleDrawer } from 'components/v2/shared/FundingCycleConfigurationDrawers/useFundingCycleDrawer'
 import UnsavedChangesModal from 'components/v2/shared/UnsavedChangesModal'
-
-import TokenForm from '../../../../pages/create/forms/TokenForm'
+import TokenForm from 'pages/create/forms/TokenForm'
 
 export default function TokenDrawer({
   visible,
