+++ src/pages/create/forms/FundingForm/DistributionSplitsSection/PayoutConfigurationExplainerCollapse.tsx
