+++ src/pages/create/forms/FundingForm/DistributionSplitsSection/index.tsx
@@ -3,7 +3,7 @@ import { t, Trans } from '@lingui/macro'
 import { Button, Form, Radio, Space } from 'antd'
 import { FormItemExt } from 'components/formItems/formItemExt'
 import TooltipIcon from 'components/TooltipIcon'
-import DistributionLimit from 'components/v2/V2Project/DistributionLimit'
+import DistributionLimit from 'components/v2/shared/DistributionLimit'
 import { CurrencyName } from 'constants/currency'
 import { ThemeContext } from 'contexts/themeContext'
 import { useWallet } from 'hooks/Wallet'
